[NOT WORKING VERSION] This update, I've been working on off and on over the past 4 months. There's over 53 file changes, and I don't even know how many lines of code was changed in this update, there's a lot.
This is a commit to start keeping track of commits and updates for vermeer since we are nearing a working release. This version is not a working release, the BrowserChangeButton is currently broken and there are things that has to be worked on. Here are a few big changes

Installer is a completely new project
Added BrowserChangeButton
Updated UniversalServer to UniversalServer 3, same thing with UniversalClient. Updated NetworkManager with that. We've also added a first error page for ssl certificates.

With that being said there's no really way to track this update time wise sincei ts so big. Much more smaller updates are on the way since we are closing in on Vermeer's release date.